fix: resolve safety/FFI findings from deep code review
Memory-safety (P0):
- cm: retain the +0 CMBlockBuffer in data_buffer_local() before wrapping,
  matching apple-cf's ownership convention (fixes over-release/use-after-free
  reachable from safe code)
- metal: give MetalDevice an ownership flag and add a real metal_device_retain
  thunk so from_ptr (borrow) / from_ptr_retained (own) contracts hold; Drop only
  releases when owned
- metal: return owned strdup copies from metal_device_get_name and the shader
  compile-error string, freed via a new metal_string_free thunk (no dangling
  NSString.utf8String)

Correctness/leaks (P1):
- config: drop the leaky pointer-keyed background-color cache; read back through
  the lifecycle-correct Swift getter (now returns a bool for set/unset)
- recording_output: make the delegate-registry locks poison-tolerant
- metal: non-panicking Clone for MetalTexture; null-check before retain in
  MetalDrawable::texture
- stream: CMClock thunk returns +0 (passUnretained) to balance from_raw's retain

Completeness/gating (P2):
- screenshot: #if-gate the macOS-14 thunks with a runtime if #available guard +
  stub (prevents a macOS-13 crash); correct the build.rs macro comment

Ergonomics (P3):
- screenshot: with_file_path/with_content_type ignore NUL input instead of
  panicking (matches existing convention)
- example 16: graceful preflight checks in main() instead of unwrap/panic

Docs/tests/tooling (P4):
- add SAFETY comments to all unsafe Send/Sync impls and stream context derefs
- add doc(cfg) to feature-gated modules; align dirty_rects count type (Int->UInt)
- add AudioBuffer repr(C) layout test, MetalTexture clone + soak tests, and an
  ignored assertive async frame-delivery test
- examples README (+24), MIGRATION note, dedup a profiling script's main()
- add scripts/check_ffi_symbols.sh (standalone Rust-extern vs Swift @_cdecl diff)

Verified: build, clippy --all-targets --all-features, full test suite (0
failures), and 151 doctests all pass.

+++ swift-bridge/Sources/CoreMedia/CoreMedia.swift
@@ -166,7 +166,7 @@ public func cm_sample_buffer_get_presenter_overlay_content_rect(_ sampleBuffer: 
 }
 
 @_cdecl("cm_sample_buffer_get_dirty_rects")
-public func cm_sample_buffer_get_dirty_rects(_ sampleBuffer: UnsafeMutableRawPointer, _ outRects: UnsafeMutablePointer<UnsafeMutableRawPointer?>, _ outCount: UnsafeMutablePointer<Int>) -> Bool {
+public func cm_sample_buffer_get_dirty_rects(_ sampleBuffer: UnsafeMutableRawPointer, _ outRects: UnsafeMutablePointer<UnsafeMutableRawPointer?>, _ outCount: UnsafeMutablePointer<UInt>) -> Bool {
     let buffer = Unmanaged<CMSampleBuffer>.fromOpaque(sampleBuffer).takeUnretainedValue()
 
     guard let attachments = CMSampleBufferGetSampleAttachmentsArray(buffer, createIfNecessary: false) as? [[CFString: Any]],
@@ -203,7 +203,7 @@ public func cm_sample_buffer_get_dirty_rects(_ sampleBuffer: UnsafeMutableRawPoi
     }
 
     outRects.pointee = UnsafeMutableRawPointer(rectsPtr)
-    outCount.pointee = rects.count
+    outCount.pointee = UInt(rects.count)
     return true
 }
